39-69 изменено место запроса для лидерборда
