+++ packages/client/src/pages/Game/hooks/useGameLogic.ts
@@ -9,9 +9,6 @@ import {
 import useLoadImages from './useLoadImages'
 import { useCanvasElements } from './useCanvasElements'
 import { ILevel } from '../constants'
-import { useAppSelector } from '@/lib/hooks/redux'
-import LeaderboardApi from '@/core/api/leaderBord.api'
-import usePlayerState from './usePlayerState'
 
 export type Modals = {
   isStartModalVisible: boolean
@@ -45,13 +42,13 @@ const useGameLogic = ({
   sounds,
   modals,
 }: GameLogicProps) => {
-  //   // Стейт для позиции главного героя
-  //   const [direction, setDirection] = useState<
-  //     'left' | 'right'
-  //   >('right')
-  //   // Стейты игрока, врагов и кристаллов
-  //   const [playerPosition, setPlayerPosition] =
-  //     useState(level.player.startPosition)
+  // Стейт для позиции главного героя
+  const [direction, setDirection] = useState<
+    'left' | 'right'
+  >('right')
+  // Стейты игрока, врагов и кристаллов
+  const [playerPosition, setPlayerPosition] =
+    useState(level.player.startPosition)
   const [gems, setGems] = useState<
     { x: number; y: number }[]
   >(level.gems.startPositions)
@@ -89,12 +86,6 @@ const useGameLogic = ({
 
   const { imagesLoaded, images } = useLoadImages()
   const {
-    playerPosition,
-    direction,
-    updatePlayerPosition,
-    setPlayerPosition,
-  } = usePlayerState(level)
-  const {
     drawObstacles,
     drawGems,
     drawPlayer,
@@ -104,10 +95,6 @@ const useGameLogic = ({
     right: images.pandaFrames,
     left: images.pandaFramesLeft,
   }
-  const leaderboardApi = new LeaderboardApi()
-  const userLogin = useAppSelector(
-    state => state.user.login
-  )
   const resetPositions = useCallback(() => {
     if (timerRef.current) {
       clearInterval(timerRef.current)
@@ -209,15 +196,11 @@ const useGameLogic = ({
           break
         case 'ArrowLeft':
           newX -= step
-          updatePlayerPosition(newX, newY, 'left')
+          setDirection('left')
           break
         case 'ArrowRight':
           newX += step
-          updatePlayerPosition(
-            newX,
-            newY,
-            'right'
-          )
+          setDirection('right')
           break
       }
 
@@ -450,18 +433,8 @@ const useGameLogic = ({
     images.foxFrames,
     direction,
   ])
-  const submitScore = useCallback(() => {
-    leaderboardApi
-      .submitScore(userLogin, score)
-      .catch(error => {
-        console.error(
-          'Error submitting score:',
-          error
-        )
-      })
-  }, [score, userLogin])
+
   const handleVictory = useCallback(() => {
-    submitScore()
     modals.showGameWinModal()
 
     modals.setIsGameActive(false)
@@ -473,7 +446,6 @@ const useGameLogic = ({
   }, [modals, sounds])
 
   const handleDefeat = useCallback(() => {
-    submitScore()
     modals.showGameOverModal()
     modals.setIsGameActive(false)
     sounds.stopGameSound()
